+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * the `deprecationWarning()` function has been replaced by the ominimum provided by the `cakephp/core` package. Be
   careful, because the signature is different now. The `TestCase::deprecated()` method has been removed and is not in
   use for now;
+* the `TestCase` class had been deprecated and were removed. Use instead the `PHPUnit\Framework\TestCase` class (and
+  possibly `ReflectionTrait` and `TestTrait`). The `deprecated()` method has been moved to `TestTrait`;
 * `Entity` and `Exceptionist` classes had been deprecated and were removed;
 * `is_json()`, `objects_map()` and `which()` global functions had been deprecated and were removed;
 * all exception classes provided by this package had been deprecated and were removed;